floating-point

+++ RenameThisClass.java
@@ -9,7 +9,7 @@ public class RenameThisClass
 {
     public static void main(String[] args) {
         int i;
-        double f = 2.5;
+        float f = 2.5f;
         boolean flag = true;
     }
 }
